requirements.txt
dash
dash-bootstrap-components
redis
python-socketio
eventlet